refactor(results dataset): join two modules related to results processing
